requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
replicate
python-dotenv==1.0.0
gunicorn==21.2.0